web/requirements.txt
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
alembic==1.7.3
Flask==2.0.1
Flask_Login==0.5.0
Flask_Migrate==3.1.0
Flask_WTF==0.15.1
Pillow==8.4.0
requests==2.26.0
SQLAlchemy==1.4.25
Werkzeug==2.0.1
WTForms==2.3.3
Flask-SSLify==0.1.5
